% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.r
\name{docx_describe_cmnts}
\alias{docx_describe_cmnts}
\title{Returns information about the comments in the Word document}
\usage{
docx_describe_cmnts(docx)
}
\arguments{
\item{docx}{\code{docx} object read with \code{read_docx}}
}
\description{
Returns information about the comments in the Word document
}
\examples{
cmnts <- read_docx(system.file("examples/comments.docx", package="docxtractr"))
docx_cmnt_count(cmnts)
docx_describe_cmnts(cmnts)
}
