% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.r
\name{docx_describe_tbls}
\alias{docx_describe_tbls}
\title{Returns a description of all the tables in the Word document}
\usage{
docx_describe_tbls(docx)
}
\arguments{
\item{docx}{\code{docx} object read with \code{read_docx}}
}
\description{
This function will attempt to discern the structure of each of the tables
in \code{docx} and print this information
}
\examples{
complx <- read_docx(system.file("examples/complex.docx", package="docxtractr"))
docx_tbl_count(complx)
docx_describe_tbls(complx)
}

