% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_find_tbls.r
\name{docx_tbl_count}
\alias{docx_tbl_count}
\title{Get number of tables in a Word document}
\usage{
docx_tbl_count(docx)
}
\arguments{
\item{docx}{\code{docx} object read with \code{read_docx}}
}
\value{
numeric
}
\description{
Get number of tables in a Word document
}
\examples{
complx <- read_docx(system.file("examples/complex.docx", package="docxtractr"))
docx_tbl_count(complx)
}
