% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/docxtractr-package.r
\docType{package}
\name{docxtractr}
\alias{docxtractr}
\alias{docxtractr-package}
\title{docxtractr is an R pacakge for extracting tables out of Word documents (docx)}
\description{
Microsoft Word docx files provide an XML structure that is fairly
straightforward to navigate, especially when it applies to Word tables. The
docxtractr package provides tools to determine table count + table structure and
extract tables from Microsoft Word docx documents.
}
\author{
Bob Rudis (@hrbrmstr)
}

