% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcga.r
\name{mcga}
\alias{mcga}
\title{Make Column Names Great Again}
\usage{
mcga(tbl)
}
\arguments{
\item{tbl}{a \code{data.frame}-like object}
}
\value{
whatever class \code{x} was but with truly great, really great column names. They're amazing.
Trust me. They'll be incredible column names once we're done.
}
\description{
Remove punctuation and spaces and turn them to underscores plus convert to lower case.
}
\examples{
real_world <- read_docx(system.file("examples/realworld.docx", package="docxtractr"))
tbls <- docx_extract_all_tbls(real_world)
mcga(assign_colnames(tbls[[1]], 2))
}
