% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.r
\name{print.docx}
\alias{print.docx}
\title{Display information about the document}
\usage{
\method{print}{docx}(x, ...)
}
\arguments{
\item{x}{\code{docx} object}

\item{...}{ignored}
}
\description{
Display information about the document
}
