% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_docs.r
\name{read_docx}
\alias{read_docx}
\title{Read in a Word document for table extraction}
\usage{
read_docx(path)
}
\arguments{
\item{path}{path to the Word document}
}
\description{
Path must be local (i.e. not a URL)
}
\examples{
doc <- read_docx(system.file("examples/data.docx", package="docxtractr"))
class(doc)
}

