% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{set_libreoffice_path}
\alias{set_libreoffice_path}
\title{Point to Local soffice.exe File}
\usage{
set_libreoffice_path(path)
}
\arguments{
\item{path}{path to the LibreOffice soffice file}
}
\value{
Returns nothing, function sets the option variable
 \code{path_to_libreoffice}.
}
\description{
Function to set an option that points to the local LibreOffice file
\code{soffice.exe}.
}
\details{
For a list of possible file path locations for \code{soffice.exe},
 see \url{https://github.com/hrbrmstr/docxtractr/issues/5#issuecomment-233181976}
}
\examples{
\dontrun{
set_libreoffice_path("local/path/to/soffice.exe")
}
}
