% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docxtractr-package.r
\docType{package}
\name{docxtractr}
\alias{docxtractr}
\alias{docxtractr-package}
\title{docxtractr is an R package for extracting tables and comments out of Word documents (docx)}
\description{
Microsoft Word docx files provide an XML structure that is fairly
straightforward to navigate, especially when it applies to Word tables. The
docxtractr package provides tools to determine table count + table structure and
extract tables from Microsoft Word docx documents. It also provides tools to determine
comment count and extract comments from Word docx documents.
}
\author{
Bob Rudis (@hrbrmstr)
}

