% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_docs.r
\name{read_docx}
\alias{read_docx}
\title{Read in a Word document for table extraction}
\usage{
read_docx(path)
}
\arguments{
\item{path}{path to the Word document}
}
\description{
Local file path or URL pointing to a \code{.docx} file. Can also take
\code{.doc} file as input if \code{LibreOffice} is installed
(see \url{https://www.libreoffice.org/} for more info and to download).
}
\examples{
doc <- read_docx(system.file("examples/data.docx", package="docxtractr"))
class(doc)
\dontrun{
# from a URL
}
}
